Move extraction functions to a utils file
