+++ src/webchat-ui/components/presentational/ScreenReaderLiveRegion.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useMemo, useState } from "react";
 import { useSelector } from "../../../webchat/helper/useSelector";
+import { cleanUpText, extractTextForScreenReader } from "../../utils/live-region-announcement";
 
 interface ScreenReaderLiveRegionProps {
 	liveContent: Record<string, string>;
@@ -22,73 +23,6 @@ const ScreenReaderLiveRegion: React.FC<ScreenReaderLiveRegionProps> = ({ liveCon
 		}
 	}, [messages, debouncedMessageIds]);
 
-	// Extract text from the DOM for screen reader by walking the DOM tree and concatenating text content while ignoring hidden elements
-	const extractTextForScreenReader = (root: HTMLElement): string => {
-		// Helper function to recursively walk through child nodes
-		const walk = (node: Node): string => {
-			if (node.nodeType === Node.TEXT_NODE) {
-				return node.textContent ?? "";
-			}
-
-			if (node.nodeType !== Node.ELEMENT_NODE) return "";
-
-			const el = node as HTMLElement;
-
-			// Skip hidden or presentational elements
-			const isHidden =
-				el.hasAttribute("aria-hidden") ||
-				el.getAttribute("role") === "presentation" ||
-				el.getAttribute("role") === "none" ||
-				el.hasAttribute("hidden") ||
-				getComputedStyle(el).display === "none" ||
-				getComputedStyle(el).visibility === "hidden";
-
-			if (isHidden) return "";
-
-			// Use aria-label if available
-			if (el.hasAttribute("aria-label")) {
-				return el.getAttribute("aria-label") + "\n";
-			}
-
-			// Handle specific elements
-			const tagHandlers: Record<string, () => string> = {
-				H1: () => walkChildren(el).trim() + ".\n",
-				H2: () => walkChildren(el).trim() + ".\n",
-				H3: () => walkChildren(el).trim() + ".\n",
-				H4: () => walkChildren(el).trim() + ".\n",
-				H5: () => walkChildren(el).trim() + ".\n",
-				H6: () => walkChildren(el).trim() + ".\n",
-				P: () => walkChildren(el).trim() + "\n",
-				LI: () => walkChildren(el).trim() + ", ",
-				UL: () => walkChildren(el),
-				OL: () => walkChildren(el),
-				BR: () => "\n",
-			};
-
-			return tagHandlers[el.tagName]?.() ?? walkChildren(el);
-		};
-
-		const walkChildren = (el: HTMLElement): string => {
-			return Array.from(el.childNodes).map(walk).join("");
-		};
-
-		return walk(root);
-	};
-
-	// Clean up text by removing emojis, HTML tags, and extra spaces
-	const cleanUpText = (text: string) => {
-		const textWithoutEmoji = text
-			.replace(
-				/([\u2700-\u27BF]|[\uE000-\uF8FF]|[\uD83C-\uDBFF\uDC00-\uDFFF]|\uFE0F|\u200D)+/gu,
-				"",
-			)
-			.replace(/<\/?[^>]+(>|$)/g, "")
-			.replace(/&nbsp;/g, " ")
-			.replace(/\s+/g, " ")
-			.trim();
-		return textWithoutEmoji;
-	};
-
 	// If live content is available for a message, use it. Otherwise, extract text from the DOM.
 	const liveText = useMemo(() => {
 		return debouncedMessageIds.map(messageId => {
@@ -99,8 +33,7 @@ const ScreenReaderLiveRegion: React.FC<ScreenReaderLiveRegionProps> = ({ liveCon
 
 			const messageElement = document.querySelector(`[data-message-id="${messageId}"]`);
 			if (messageElement) {
-				const extractedText = extractTextForScreenReader(messageElement as HTMLElement);
-				const cleanText = cleanUpText(extractedText);
+				const cleanText = extractTextForScreenReader(messageElement as HTMLElement);
 				return <div key={messageId}>{cleanText || "A new message"}</div>;
 			}
